requirements.txt
facenet-pytorch>=2.5.3
Pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
scikit-learn